Move events into top level module
