+++ include/BLIB/Events.hpp
@@ -1,0 +1,8 @@
+#ifndef BLIB_EVENTS_HPP
+#define BLIB_EVENTS_HPP
+
+#include <BLIB/Events/EventDispatcher.hpp>
+#include <BLIB/Events/EventDispatcherScopeGuard.hpp>
+#include <BLIB/Events/EventListener.hpp>
+
+#endif
