+++ include/BLIB/Events/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(BLIB PUBLIC
+    EventDispatcher.hpp
+    EventDispatcherScopeGuard.hpp
+    EventListener.hpp
+)
