+++ include/BLIB/Util.hpp
@@ -6,9 +6,6 @@
  * @brief Miscellaneous set of classes that provide common functionality
  */
 
-#include <BLIB/Util/EventDispatcher.hpp>
-#include <BLIB/Util/EventDispatcherScopeGuard.hpp>
-#include <BLIB/Util/EventListener.hpp>
 #include <BLIB/Util/NonCopyable.hpp>
 #include <BLIB/Util/PairHash.hpp>
 #include <BLIB/Util/Random.hpp>
