+++ tests/Events/CMakeLists.txt
@@ -1,3 +1,3 @@
 target_sources(BLIB.t PUBLIC
-    Signal.t.cpp
+    EventDispatcher.t.cpp
 )
